refactor: move get all specialization files into getall directory
