+++ apps/api/core/users-jobs-service/src/Application/Specializations/GetAll/GetAllQuery.cs
@@ -1,4 +1,4 @@
-namespace Application.Specializations;
+namespace Application.Specializations.GetAll;
 
 using SkApplication.Contracts;
