+++ apps/api/core/users-jobs-service/src/Application/Specializations/GetAll/GetAllQueryHandler.cs
@@ -1,8 +1,8 @@
-namespace Application.Specializations;
+namespace Application.Specializations.GetAll;
 
+using Application.Persistent;
 using Domain.Attributes.Specializations;
 using Microsoft.EntityFrameworkCore;
-using Persistent;
 using SkApplication.Contracts;
 using SkDomain.Results;
 
@@ -18,14 +18,13 @@ internal sealed class GetAllQueryHandler(IApplicationDbContext context)
         IQueryable<Specialization> specializations =
             context.Specializations.OrderBy(specialization => specialization.Name);
 
-        List<Specialization> specializationsList =
-            await specializations.ToListAsync(cancellationToken);
-
-
         if (!await specializations.AnyAsync(cancellationToken))
         {
             return Result.Failure<Response>(SpecializationErrors.NotSpecializationsFound);
         }
+
+        List<Specialization> specializationsList =
+            await specializations.ToListAsync(cancellationToken);
 
         Response response = converter.Convert(specializationsList);
         return Result.Success(response);
