+++ apps/api/core/users-jobs-service/src/Application/Specializations/GetAll/Validator.cs
@@ -1,4 +1,4 @@
-namespace Application.Specializations;
+namespace Application.Specializations.GetAll;
 
 using FluentValidation;
